Simplify theming for `Rule` widget
