+++ style/src/lib.rs
@@ -20,7 +20,6 @@ pub mod application;
 pub mod menu;
 pub mod pane_grid;
 pub mod pick_list;
-pub mod rule;
 pub mod svg;
 pub mod text_editor;
 pub mod theme;
